move icurrentuser and mappers to auth-module
