+++ apps/server/src/modules/account/controller/account.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Delete, Get, Param, Patch, Query } from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { Authenticate, CurrentUser } from '@src/modules/authentication/decorator/auth.decorator';
-import { ICurrentUser } from '@shared/domain';
 import { EntityNotFoundError, ForbiddenOperationError, ValidationError } from '@shared/common';
+import { ICurrentUser } from '../../authentication';
 import { AccountUc } from '../uc/account.uc';
 import {
 	AccountByIdBodyParams,
