+++ apps/server/src/modules/authentication/controllers/login.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Req, Post, UseGuards, HttpCode, HttpStatus } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
-import { ICurrentUser } from '@shared/domain';
+import { ICurrentUser } from '@src/modules/authentication';;
 import { ApiTags } from '@nestjs/swagger';
 import { AuthenticationService } from '../services/authentication.service';
