+++ apps/server/src/modules/authentication/index.ts
@@ -1,1 +1,3 @@
 export * from './authentication.module';
+export * from './interface';
+export * from './mapper';
