+++ apps/server/src/modules/authentication/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './user';
