+++ apps/server/src/modules/authentication/mapper/current-user.mapper.spec.ts
