+++ apps/server/src/modules/authentication/mapper/index.ts
@@ -1,0 +1,1 @@
+export * from './current-user.mapper';
