+++ apps/server/src/modules/authentication/strategy/local.strategy.ts
@@ -4,11 +4,11 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import bcrypt from 'bcryptjs';
 import { UserRepo } from '@shared/repo';
-import { ICurrentUser } from '@shared/domain';
-import { CurrentUserMapper } from '@shared/domain/mapper/current-user.mapper';
+import { CurrentUserMapper } from '@shared/../mapper/current-user.mapper';
 import { AccountDto } from '@src/modules/account/services/dto';
 import { GuardAgainst } from '@shared/common/utils/guard-against';
 import { IdentityManagementOauthService, IIdentityManagementConfig } from '@shared/infra/identity-management';
+import { ICurrentUser } from '../interface';
 import { AuthenticationService } from '../services/authentication.service';
 
 @Injectable()
