+++ apps/server/src/modules/collaborative-storage/controller/collaborative-storage.controller.spec.ts
@@ -2,7 +2,7 @@ import { CollaborativeStorageController } from '@src/modules/collaborative-stora
 import { Test, TestingModule } from '@nestjs/testing';
 import { CollaborativeStorageUc } from '@src/modules/collaborative-storage/uc/collaborative-storage.uc';
 import { createMock } from '@golevelup/ts-jest';
-import { ICurrentUser } from '@shared/domain';
+import { ICurrentUser } from '@src/modules/authentication';;
 import { Logger } from '@src/core/logger';
 
 describe('CollaborativeStorage Controller', () => {
