+++ apps/server/src/modules/learnroom/controller/cards.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Query } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
-import { ICurrentUser } from '@shared/domain';
+import { ICurrentUser } from '@src/modules/authentication';;
 import { Authenticate, CurrentUser } from '@src/modules/authentication/decorator/auth.decorator';
 import { CardIdsParams, CardListResponse, CardResponse, ContentElementResponse } from './dto';
 import { VisibilitySettingsResponse } from './dto/card/visibility-settings.response';
