+++ apps/server/src/modules/lesson/controller/lesson.controller.spec.ts
@@ -1,6 +1,6 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { Test, TestingModule } from '@nestjs/testing';
-import { ICurrentUser } from '@shared/domain';
+import { ICurrentUser } from '@src/modules/authentication';;
 import { LessonUC } from '../uc';
 import { LessonController } from './lesson.controller';
