+++ apps/server/src/modules/lesson/controller/lesson.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Delete, Param } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
-import { ICurrentUser } from '@shared/domain';
+import { ICurrentUser } from '@src/modules/authentication';;
 import { Authenticate, CurrentUser } from '@src/modules/authentication/decorator/auth.decorator';
 import { LessonUC } from '../uc';
 import { LessonUrlParams } from './dto';
