+++ apps/server/src/modules/task/controller/task.controller.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from '@golevelup/ts-jest';
 import { Test, TestingModule } from '@nestjs/testing';
-import { ICurrentUser } from '@shared/domain';
+import { ICurrentUser } from '@src/modules/authentication';;
 import { CopyElementType, CopyStatus, CopyStatusEnum } from '@src/modules/copy-helper';
 import { CopyApiResponse } from '@src/modules/copy-helper/dto/copy.response';
 import { TaskCopyUC, TaskUC } from '../uc';
