+++ apps/server/src/shared/common/interceptor/request-logging.interceptor.ts
@@ -1,9 +1,9 @@
 import { CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nestjs/common';
-import { ICurrentUser } from '@shared/domain';
 import { Logger, RequestLoggingBody } from '@src/core/logger';
 import { Request } from 'express';
 import { Observable, throwError } from 'rxjs';
 import { catchError, tap } from 'rxjs/operators';
+import { ICurrentUser } from '../../../modules';
 
 @Injectable()
 export class RequestLoggingInterceptor implements NestInterceptor {
