+++ apps/server/src/shared/domain/interface/index.ts
@@ -5,5 +5,4 @@ export * from './learnroom';
 export * from './permission';
 export * from './permission.enum';
 export * from './rolename.enum';
-export * from './user';
 export * from './video-conference-scope.enum';
